# implementation/virt-reader/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AJIN RFID Tracking API",
        "description": "RFID 기반 물류 추적 시스템 API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/pallets": {
            "post": {
                "tags": [
                    "Pallets"
                ],
                "summary": "Create Pallet",
                "description": "팔레트 생성 (권한: pallets:write). rfid_epc 제공 시 실물 팔레트 자동 생성.",
                "operationId": "create_pallet_api_v1_pallets_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PalletCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PalletResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Pallets"
                ],
                "summary": "List Pallets",
                "description": "팔레트 목록 조회 (권한: pallets:read)",
                "operationId": "list_pallets_api_v1_pallets_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 20,
                            "title": "Per Page"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    },
                    {
                        "name": "lot_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Lot Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PalletListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/physical-pallets": {
            "post": {
                "tags": [
                    "Physical Pallets"
                ],
                "summary": "Create Physical Pallet",
                "description": "실물 팔레트 생성 (권한: physical_pallets:write)",
                "operationId": "create_physical_pallet_api_v1_physical_pallets_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PhysicalPalletCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhysicalPalletResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Physical Pallets"
                ],
                "summary": "List Physical Pallets",
                "description": "실물 팔레트 목록 조회 (권한: physical_pallets:read)",
                "operationId": "list_physical_pallets_api_v1_physical_pallets_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "description": "페이지 번호",
                            "default": 1,
                            "title": "Page"
                        },
                        "description": "페이지 번호"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "페이지당 항목 수",
                            "default": 20,
                            "title": "Per Page"
                        },
                        "description": "페이지당 항목 수"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "상태 필터",
                            "title": "Status"
                        },
                        "description": "상태 필터"
                    },
                    {
                        "name": "item_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "품목 ID 필터",
                            "title": "Item Id"
                        },
                        "description": "품목 ID 필터"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "검색어 (EPC, 팔레트 코드)",
                            "title": "Search"
                        },
                        "description": "검색어 (EPC, 팔레트 코드)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhysicalPalletListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/lots/receiving": {
            "post": {
                "tags": [
                    "Lots"
                ],
                "summary": "Create Receiving Lot",
                "description": "원자재 입고 LOT 생성 (RFID 불필요, 수동 등록) (권한: lots:write)",
                "operationId": "create_receiving_lot_api_v1_lots_receiving_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LotReceiving"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LotResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PalletCreate": {
                "properties": {
                    "pallet_no": {
                        "type": "string",
                        "title": "Pallet No",
                        "description": "팔레트 번호"
                    },
                    "physical_pallet_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Physical Pallet Id",
                        "description": "실물 팔레트 ID"
                    },
                    "quantity": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Quantity",
                        "description": "현재 적재 수량",
                        "default": 0
                    },
                    "rfid_epc": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rfid Epc",
                        "description": "RFID EPC (제공 시 PhysicalPallet 생성 후 연결)"
                    },
                    "status": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Status",
                        "description": "초기 상태 (Empty, Stock 등)"
                    }
                },
                "type": "object",
                "required": [
                    "pallet_no"
                ],
                "title": "PalletCreate",
                "description": "팔레트 생성 시 rfid_epc 제공 시 실물 팔레트 자동 생성"
            },
...
            "PalletListResponse": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/PalletResponse"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "per_page": {
                        "type": "integer",
                        "title": "Per Page"
                    },
                    "pages": {
                        "type": "integer",
                        "title": "Pages"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "total",
                    "page",
                    "per_page",
                    "pages"
                ],
                "title": "PalletListResponse"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}